+++ lib/govuk_app_config/govuk_healthcheck/checkup.rb
@@ -46,9 +46,13 @@ module GovukHealthcheck
       if check.respond_to?(:enabled?) && !check.enabled?
         { status: :ok, message: "currently disabled" }
       else
-        component_status = details(check).merge(status: check.status)
-        component_status[:message] = check.message if check.respond_to?(:message)
-        component_status
+        begin
+          component_status = details(check).merge(status: check.status)
+          component_status[:message] = check.message if check.respond_to?(:message)
+          component_status
+        rescue StandardError => e
+          { status: :critical, message: e.to_s }
+        end
       end
     end